[Refactor] 세팅 페이지 파일 구조 변경 및 뷰모델 분리
